+++ internal/processing/synchronous/status/get.go
