+++ internal/timeline/manager.go
